src/normalizer/requirements.txt
PyYAML==6.0
pandas==1.5.3
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
psycopg2==2.9.5